fishapi/requirements.txt
Django==5.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
tensorflow>=2.15.0
keras>=2.15.0
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0

# Chatbot dependencies
python-dotenv>=1.0.1
rdflib>=7.1.3
openai>=1.107.1
deepseek>=1.0.0